+++ backup/__init__.py
@@ -1,0 +1,3 @@
+from .backup import LogBackupThread
+
+__all__ = ['LogBackupThread']