+++ ui/components/settings_dialog.py
@@ -1,5 +1,4 @@
-from PyQt5.QtCore import Qt
-from PyQt5.QtCore import QDateTime, pyqtSignal
+from PyQt5.QtCore import pyqtSignal, QDateTime, Qt
 from PyQt5.QtWidgets import QDialog, QFileDialog
 
 import os
@@ -9,6 +8,7 @@ from ..mixins import FormMixin
 from ..tasks import AsynchronousTask
 from ..forms import Form, Field
 from ..forms.validators import not_empty
+
 
 class SettingsDialog(QDialog, FormMixin):
     form_class = Ui_SettingsDialog
@@ -25,6 +25,9 @@ class SettingsDialog(QDialog, FormMixin):
         active_tool = kwargs.pop('active_tool')
         logs_location = kwargs.pop('logs_location')
         logs_date_start = kwargs.pop('logs_date_start')
+        enable_logs_backup = kwargs.pop('enable_logs_backup')
+        logs_backup_location = kwargs.pop('logs_backup_location')
+        logs_backup_action = kwargs.pop('logs_backup_action')
 
         super().__init__(*args, **kwargs)
 
@@ -46,6 +49,11 @@ class SettingsDialog(QDialog, FormMixin):
                 logs_date_start.second,
             )
         )
+        if enable_logs_backup:
+            self.form.checkbox_activate_backup.setChecked(True)
+            self.form.line_backup_location.setText(logs_backup_location)
+            self.form.combo_backup_action.setCurrentText(logs_backup_action)
+
         self.__settings_form = Form(
             fields=(
                 Field(name='Endereço URL', widget=self.form.line_cp_url, validators=(not_empty,)),
@@ -53,6 +61,9 @@ class SettingsDialog(QDialog, FormMixin):
                 Field(name='Diretório', widget=self.form.line_logs_location, validators=(not_empty,)),
                 Field(name='Ferramenta', widget=self.form.combo_logs_tool, validators=(not_empty,)),
                 Field(name='Apartir de:', widget=self.form.line_logs_date_start),
+                # Field(name='Ativar backup automático', widget=self.form.checkbox_activate_backup),
+                Field(name='Diretório', widget=self.form.line_backup_location),
+                Field(name='Ação', widget=self.form.combo_backup_action)
             ),
             on_submit=self.__save_handler,
             submit_button=self.form.btn_save,
@@ -62,7 +73,12 @@ class SettingsDialog(QDialog, FormMixin):
         self.__get_workstation_list()
 
     def set_bindings(self):
-        self.form.btn_dir_explorer.clicked.connect(self.__open_dir_explorer_handler)
+        self.form.btn_dir_explorer.clicked.connect(
+            lambda _: self.__open_dir_explorer_handler(self.form.line_logs_location)
+        )
+        self.form.btn_dir_explorer_backup.clicked.connect(
+            lambda _: self.__open_dir_explorer_handler(self.form.line_backup_location),
+        )
         self.form.btn_update_workstations.clicked.connect(self.__get_workstation_list)
         self.workstation_list_collected_signal.connect(self.__update_workstation_list_options)
 
@@ -72,17 +88,17 @@ class SettingsDialog(QDialog, FormMixin):
 
         event.accept()
 
-    def __open_dir_explorer_handler(self):
+    def __open_dir_explorer_handler(self, widget):
         dialog = QFileDialog()
 
         new_path = os.path.normpath(dialog.getExistingDirectory(
             self,
             'Selecionar localização dos logs',
-            self.form.line_logs_location.text()
+            widget.text()
         ))
 
         if new_path != '.':
-            self.form.line_logs_location.setText(new_path)
+            widget.setText(new_path)
 
     def __save_handler(self):
         new_settings = {
@@ -90,7 +106,10 @@ class SettingsDialog(QDialog, FormMixin):
             'active_tool': self.form.combo_logs_tool.currentText(),
             'active_workstation': self.form.combo_cp_workstation.currentText(),
             'logs_location': self.form.line_logs_location.text(),
-            'logs_date_start': self.form.line_logs_date_start.dateTime().toPyDateTime()
+            'logs_date_start': self.form.line_logs_date_start.dateTime().toPyDateTime(),
+            'enable_logs_backup': self.form.checkbox_activate_backup.isChecked(),
+            'logs_backup_location': self.form.line_backup_location.text(),
+            'logs_backup_action': self.form.combo_backup_action.currentText()
         }
 
         self.on_save(new_settings)
@@ -115,4 +134,3 @@ class SettingsDialog(QDialog, FormMixin):
 
         self.form.btn_update_workstations.setEnabled(False)
         self.task.start()
-