+++ worker/__init__.py
@@ -1,4 +1,5 @@
 from .worker import Worker
 from .parser import Parser, TOOLS
 
+
 __all__ = ['Worker', 'Parser', 'TOOLS']